+++ src/App.css
@@ -20,12 +20,12 @@
     transform: rotate(360deg);
   }
 }
-header svg{
+.header svg{
   fill: #ffe81f;
   height: 75px;
 
 }
-header{
+.header{
   text-align: center;
   margin-top: 30px;
   margin-bottom: 15px;
